pyproject.toml
[project]
name = "inthememory"
version = "0.1.0"
description = ""
authors = [
    {name = "baptprr",email = "baptiste_poirier@yahoo.fr"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pandas (>=2.3.1,<3.0.0)",
    "azure-storage-blob (>=12.25.1,<13.0.0)",
    "pyarrow (>=20.0.0,<21.0.0)",
    "pytest (>=8.4.1,<9.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
